move shared tests data to a helpers file
