+++ tests/client/amo/containers/TestDetail.js
@@ -9,9 +9,8 @@ import { Provider } from 'react-redux';
 import createStore from 'amo/store';
 import DetailPage from 'amo/containers/DetailPage';
 import I18nProvider from 'core/i18n/Provider';
+import { fakeAddon } from 'tests/client/amo/helpers';
 import { getFakeI18nInst } from 'tests/client/helpers';
-
-import { fakeAddon } from '../components/TestAddonDetail';
 
 function render(
   {
